+++ src/zeus2_toolbox/zeus2_io.py
@@ -12,6 +12,7 @@ import configparser
 import copy
 import gc
 import inspect
+import os.path
 from collections import Counter
 from datetime import datetime, timezone
 
@@ -150,25 +151,8 @@ class TableObj(BaseObj):
             if self.empty_flag_:
                 table_new = other.table_
             else:
-                try:
-                    table_new = vstack([self.table_, other.table_],
-                                       join_type="outer")
-                except np_utils.TableMergeError as err:
-                    if "columns have incompatible types" in err.args[0]:
-                        colname = err.args[0].split("'")[1]
-                        tb_tmp, tb_other_tmp = self.table_.copy(), \
-                                               other.table_.copy()
-                        try:
-                            tb_tmp[colname] = tb_tmp[colname].astype(
-                                    tb_other_tmp[colname].dtype)
-                        except TypeError:
-                            tb_other_tmp[colname] = tb_other_tmp[colname].astype(
-                                    tb_tmp[colname].dtype)
-                        table_new = vstack([tb_tmp, tb_other_tmp],
-                                           join_type="outer")
-                        del (tb_tmp, tb_other_tmp)
-                    else:
-                        raise err
+                table_new = vstack_reconcile(self.table_, other.table_,
+                                             join_type="outer")
             self.__fill_values__(tb_in=table_new)
 
     def expand(self, other):
@@ -2185,6 +2169,8 @@ class TimeStamps(DataObj):
             will try to read filename.ts if the first attempt fails
         :return ts: TimeStamps, new object
         :rtype: TimeStamps
+        :raises RuntimeError: failed to read filename and filename.ts
+        :raises FileNotFoundError: filename and filename.ts don't exist
 
         :example:
         ts = TimeStamps.read(filename='irc10216_191128_0065.ts')
@@ -2192,16 +2178,41 @@ class TimeStamps(DataObj):
         if filename is None:
             arr_in = None
         else:
-            try:
-                arr_in = Tb.read(filename, format="ascii")["col2"]. \
-                    data.astype(np.double)
-            except FileNotFoundError:
+            if os.path.isfile(filename):
                 try:
-                    arr_in = Tb.read(filename + ".ts", format="ascii")["col2"]. \
+                    arr_in = Tb.read(filename, format="ascii.no_header",
+                                     fast_reader=True, delimiter=" ",
+                                     guess=False)["col2"]. \
                         data.astype(np.double)
-                except FileNotFoundError:
-                    raise FileNotFoundError("%s or %s.ts does not exist." %
-                                            (filename, filename))
+                except Exception as err:
+                    if os.path.isfile(filename + ".ts"):
+                        try:
+                            arr_in = Tb.read(filename + ".ts",
+                                             format="ascii.no_header",
+                                             fast_reader=True, delimiter=" ",
+                                             guess=False)["col2"]. \
+                                data.astype(np.double)
+                        except Exception as err1:
+                            raise RuntimeError(("failed to read %s and %s.ts." %
+                                                (filename, filename))) from err1
+                    else:
+                        raise RuntimeError(
+                                "failed to read %s, and %s.ts doesn't exist." %
+                                (filename, filename)) from err
+            elif os.path.isfile(filename + ".ts"):
+                try:
+                    arr_in = Tb.read(filename + ".ts",
+                                     format="ascii.no_header",
+                                     fast_reader=True, delimiter=" ",
+                                     guess=False)["col2"]. \
+                        data.astype(np.double)
+                except Exception as err2:
+                    raise RuntimeError(
+                            "failed to read %s.ts, and %s doesn't exist." %
+                            (filename, filename)) from err2
+            else:
+                raise FileNotFoundError("%s and %s.ts don't exist." %
+                                        (filename, filename))
         ts = cls(arr_in=arr_in)
         ts.check()
 
@@ -4164,6 +4175,62 @@ class ObsArray(Obs):
                         col_data, name="%s=%i" % (colname_repeat, col_repeat)))
 
         return tb
+
+
+def vstack_reconcile(tb1, tb2, **kwargs):
+    """
+    try to vstack two tables, in the case of incompatible column types, will
+    try to convert the type to a common type and try again
+
+    :param astropy.table.Table tb1: astropy.table class or TableObj instance,
+        the table to be appended
+    :type tb1: astropy.table.table.Table or TableObj
+    :param astropy.table.Table tb2: astropy.table class or TableObj instance,
+        the table to append
+    :type tb2: astropy.table.table.Table or TableObj
+    :param dict kwargs: keyword arguments passed to vstack
+    :return: stacked table
+    :rtype: astropy.table.table.Table
+    :raises TypeError: can not reconcile the type
+    """
+
+    if isinstance(tb1, TableObj):
+        tb1 = tb1.table_
+    elif not isinstance(tb1, Tb):
+        raise TypeError("Invalid input type for tb1.")
+
+    if isinstance(tb2, TableObj):
+        tb1 = tb2.table_
+    elif not isinstance(tb1, Tb):
+        raise TypeError("Invalid input type for tb2.")
+
+    try:
+        table_new = vstack([tb1, tb2], **kwargs)
+    except Exception as err:
+        if "columns have incompatible types" in err.args[0]:
+            colname = err.args[0].split("'")[1]
+            tb1_tmp, tb2_tmp = tb1.copy(), tb2.copy()
+            try:
+                tb1_tmp[colname] = tb1_tmp[colname].astype(
+                        tb2_tmp[colname].dtype)
+            except TypeError:
+                try:
+                    tb2_tmp[colname] = \
+                        tb2_tmp[colname].astype(
+                                tb1_tmp[colname].dtype)
+                except TypeError as err1:
+                    raise TypeError(
+                            "can not reconcile the type " +
+                            "%s and %s of the column %s." %
+                            (tb1_tmp[colname].dtype,
+                             tb2_tmp[colname].dtype, colname)) \
+                        from err
+            table_new = vstack_reconcile(tb1_tmp, tb2_tmp, **kwargs)
+            del (tb1_tmp, tb2_tmp)
+        else:
+            raise err
+
+    return table_new
 
 
 def check_array_type(array_type):